Initial preparations for adding Scala 3 support
