+++ shared/src/main/scala-2.13/zipper/ForImpl.scala
@@ -6,6 +6,16 @@ import shapeless.ops.hlist.{Replacer, Selector}
 import scala.collection.Factory
 
 private[zipper] trait ForImpl {
+  implicit def `Unzip List-based`[A, L <: HList](
+    implicit generic: Generic.Aux[A, L],
+    select: Selector[L, List[A]],
+    replace: Replacer.Aux[L, List[A], List[A], (List[A], L)]
+  ): Unzip[A] = new Unzip[A] {
+    def unzip(node: A): List[A] = select(generic.to(node))
+
+    def zip(node: A, children: List[A]): A = generic.from(replace(generic.to(node), children)._2)
+  }
+
   class For[A, Coll[X] <: Seq[X]] {
     /** Derive an instance of `Unzip[A]` */
     def derive[L <: HList](
